memoize fields
