+++ packages/react-client/src/hooks/devices/useMicrophone.ts
@@ -1,6 +1,6 @@
 import type { Device } from "../../types/public";
-import { useDeviceManager } from "../deviceManagers/useDeviceManager";
 import { useFishjamContext } from "../useFishjamContext";
+import { useDeviceApi } from "./useDeviceApi";
 
 /**
  *
@@ -8,33 +8,6 @@ import { useFishjamContext } from "../useFishjamContext";
  */
 export function useMicrophone(): Device {
   const { audioTrackManager, audioDeviceManagerRef } = useFishjamContext();
-  const { deviceState, status } = useDeviceManager(audioDeviceManagerRef.current);
 
-  const { currentTrack, ...trackManager } = audioTrackManager;
-
-  const stream = deviceState.media?.stream ?? null;
-  const currentMiddleware = deviceState.currentMiddleware ?? null;
-  const isStreaming = Boolean(currentTrack?.stream);
-  const track = stream?.getAudioTracks()[0] ?? null;
-  const trackId = currentTrack?.trackId ?? null;
-  const devices = deviceState.devices ?? [];
-  const activeDevice = deviceState.media?.deviceInfo ?? null;
-  const isMuted = !deviceState.media?.enabled;
-  const deviceError = deviceState.error ?? null;
-  const isDeviceEnabled = Boolean(deviceState.media);
-
-  return {
-    ...trackManager,
-    currentMiddleware,
-    isDeviceEnabled,
-    status,
-    stream,
-    track,
-    isStreaming,
-    trackId,
-    devices,
-    activeDevice,
-    isMuted,
-    deviceError,
-  };
+  return useDeviceApi({ trackManager: audioTrackManager, deviceManager: audioDeviceManagerRef.current });
 }
